Refactor GFS forecast reformatters and configs
This commit refactors the NOAA GFS forecast reformatters and configuration classes. It introduces common base classes for region job and template configuration, promoting code reuse and simplifying the structure. The changes include:

- **`src/reformatters/noaa/gfs/forecast/region_job.py`**:
    - Moves common GFS region job logic into a new base class `NoaaGfsCommonRegionJob`.
    - `NoaaGfsForecastRegionJob` now inherits from `NoaaGfsCommonRegionJob`.
    - `NoaaGfsForecastSourceFileCoord` now inherits from `NoaaGfsSourceFileCoord`.
    - Removes redundant code and improves testability by using module-local imports.

- **`src/reformatters/noaa/gfs/forecast/template_config.py`**:
    - Moves common GFS template configuration logic into a new base class `NoaaGfsCommonTemplateConfig`.
    - `NoaaGfsForecastTemplateConfig` now inherits from `NoaaGfsCommonTemplateConfig`.
    - Simplifies coordinate and data variable definitions by leveraging the common base class.

- **New files**:
    - **`src/reformatters/noaa/gfs/region_job.py`**: Contains the new `NoaaGfsCommonRegionJob` and `NoaaGfsSourceFileCoord` base classes.
    - **`src/reformatters/noaa/gfs/template_config.py`**: Contains the new `NoaaGfsCommonTemplateConfig` base class.
    - **`tests/noaa/gfs/region_job_test.py`**: Adds unit tests for the common region job functionality.
    - **`tests/noaa/gfs/template_config_test.py`**: Adds unit tests for the common template configuration functionality.

These changes aim to improve code organization, reduce duplication, and enhance the maintainability of the NOAA GFS reformatting components.
