+++ src/reformatters/noaa/gfs/template_config.py
@@ -1,0 +1,488 @@
+from collections.abc import Sequence
+
+import numpy as np
+import pandas as pd
+
+from reformatters.common.config_models import (
+    Coordinate,
+    CoordinateAttrs,
+    DataVarAttrs,
+    Encoding,
+    StatisticsApproximate,
+)
+from reformatters.common.template_config import TemplateConfig
+from reformatters.common.zarr import (
+    BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE,
+)
+from reformatters.noaa.models import NoaaDataVar, NoaaInternalAttrs
+
+
+class NoaaGfsCommonTemplateConfig(TemplateConfig[NoaaDataVar]):
+    """Common template configuration for GFS datasets. Subclassed by forecast and analysis datasets."""
+
+    def _get_latitude_coord(
+        self, lat_values: np.ndarray[tuple[int], np.dtype[np.float64]]
+    ) -> Coordinate:
+        """Return latitude coordinate config."""
+        return Coordinate(
+            name="latitude",
+            encoding=Encoding(
+                dtype="float64",
+                fill_value=np.nan,
+                compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                chunks=len(lat_values),
+                shards=None,
+            ),
+            attrs=CoordinateAttrs(
+                units="degrees_north",
+                statistics_approximate=StatisticsApproximate(
+                    min=float(lat_values.min()),
+                    max=float(lat_values.max()),
+                ),
+            ),
+        )
+
+    def _get_longitude_coord(
+        self, lon_values: np.ndarray[tuple[int], np.dtype[np.float64]]
+    ) -> Coordinate:
+        """Return longitude coordinate config."""
+        return Coordinate(
+            name="longitude",
+            encoding=Encoding(
+                dtype="float64",
+                fill_value=np.nan,
+                compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                chunks=len(lon_values),
+                shards=None,
+            ),
+            attrs=CoordinateAttrs(
+                units="degrees_east",
+                statistics_approximate=StatisticsApproximate(
+                    min=float(lon_values.min()),
+                    max=float(lon_values.max()),
+                ),
+            ),
+        )
+
+    def _get_spatial_ref_coord(self) -> Coordinate:
+        """Return spatial_ref coordinate config."""
+        return Coordinate(
+            name="spatial_ref",
+            encoding=Encoding(
+                dtype="int64",
+                fill_value=0,
+                chunks=(),  # Scalar coordinate
+                shards=None,
+            ),
+            attrs=CoordinateAttrs(
+                units=None,
+                statistics_approximate=None,
+                # Deterived by running `ds.rio.write_crs("+proj=longlat +a=6371229 +b=6371229 +no_defs +type=crs")["spatial_ref"].attrs
+                crs_wkt='GEOGCS["unknown",DATUM["unknown",SPHEROID["unknown",6371229,0]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AXIS["Longitude",EAST],AXIS["Latitude",NORTH]]',
+                semi_major_axis=6371229.0,
+                semi_minor_axis=6371229.0,
+                inverse_flattening=0.0,
+                reference_ellipsoid_name="unknown",
+                longitude_of_prime_meridian=0.0,
+                prime_meridian_name="Greenwich",
+                geographic_crs_name="unknown",
+                horizontal_datum_name="unknown",
+                grid_mapping_name="latitude_longitude",
+                spatial_ref='GEOGCS["unknown",DATUM["unknown",SPHEROID["unknown",6371229,0]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AXIS["Longitude",EAST],AXIS["Latitude",NORTH]]',
+                comment="This coordinate reference system matches the source data which follows WMO conventions of assuming the earth is a perfect sphere with a radius of 6,371,229m. It is similar to EPSG:4326, but EPSG:4326 uses a more accurate representation of the earth's shape.",
+            ),
+        )
+
+    def get_data_vars(self, encoding: Encoding) -> Sequence[NoaaDataVar]:
+        """Return the data variables for the dataset with the given encoding."""
+        default_window_reset_frequency = pd.Timedelta("6h")
+        default_keep_mantissa_bits = 7
+
+        return [
+            NoaaDataVar(
+                name="pressure_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="sp",
+                    long_name="Surface pressure",
+                    units="Pa",
+                    step_type="instant",
+                    standard_name="surface_air_pressure",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="PRES",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=560,
+                    keep_mantissa_bits=10,
+                ),
+            ),
+            NoaaDataVar(
+                name="temperature_2m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="t2m",
+                    long_name="2 metre temperature",
+                    units="C",
+                    step_type="instant",
+                    standard_name="air_temperature",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="TMP",
+                    grib_description='2[m] HTGL="Specified height level above ground"',
+                    grib_index_level="2 m above ground",
+                    index_position=580,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="relative_humidity_2m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="r2",
+                    long_name="2 metre relative humidity",
+                    units="%",
+                    step_type="instant",
+                    standard_name="relative_humidity",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="RH",
+                    grib_description='2[m] HTGL="Specified height level above ground"',
+                    grib_index_level="2 m above ground",
+                    index_position=583,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="maximum_temperature_2m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="tmax",
+                    long_name="Maximum temperature",
+                    units="C",
+                    step_type="max",
+                    comment="Maximum over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="TMAX",
+                    grib_description='2[m] HTGL="Specified height level above ground"',
+                    grib_index_level="2 m above ground",
+                    index_position=585,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="minimum_temperature_2m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="tmin",
+                    long_name="Minimum temperature",
+                    units="C",
+                    step_type="min",
+                    comment="Minimum over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="TMIN",
+                    grib_description='2[m] HTGL="Specified height level above ground"',
+                    grib_index_level="2 m above ground",
+                    index_position=586,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="wind_u_10m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="u10",
+                    long_name="10 metre U wind component",
+                    units="m/s",
+                    step_type="instant",
+                    standard_name="eastward_wind",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="UGRD",
+                    grib_description='10[m] HTGL="Specified height level above ground"',
+                    grib_index_level="10 m above ground",
+                    index_position=587,
+                    keep_mantissa_bits=6,
+                ),
+            ),
+            NoaaDataVar(
+                name="wind_v_10m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="v10",
+                    long_name="10 metre V wind component",
+                    units="m/s",
+                    step_type="instant",
+                    standard_name="northward_wind",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="VGRD",
+                    grib_description='10[m] HTGL="Specified height level above ground"',
+                    grib_index_level="10 m above ground",
+                    index_position=588,
+                    keep_mantissa_bits=6,
+                ),
+            ),
+            NoaaDataVar(
+                name="wind_u_100m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="u100",
+                    long_name="100 metre U wind component",
+                    standard_name="eastward_wind",
+                    units="m/s",
+                    step_type="instant",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="UGRD",
+                    grib_description='100[m] HTGL="Specified height level above ground"',
+                    grib_index_level="100 m above ground",
+                    index_position=688,
+                    keep_mantissa_bits=6,
+                ),
+            ),
+            NoaaDataVar(
+                name="wind_v_100m",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="v100",
+                    long_name="100 metre V wind component",
+                    units="m/s",
+                    step_type="instant",
+                    standard_name="northward_wind",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="VGRD",
+                    grib_index_level="100 m above ground",
+                    grib_description='100[m] HTGL="Specified height level above ground"',
+                    index_position=689,
+                    keep_mantissa_bits=6,
+                ),
+            ),
+            NoaaDataVar(
+                name="percent_frozen_precipitation_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="cpofp",
+                    long_name="Percent frozen precipitation",
+                    units="%",
+                    step_type="instant",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="CPOFP",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=590,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="precipitation_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="tp",
+                    long_name="Total Precipitation",
+                    units="mm/s",
+                    comment="Average precipitation rate since the previous forecast step.",
+                    step_type="avg",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="APCP",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=595,
+                    deaccumulate_to_rate=True,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="categorical_snow_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="csnow",
+                    long_name="Categorical snow",
+                    units="0=no; 1=yes",
+                    step_type="avg",
+                    comment="Presence/absence over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="CSNOW",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=604,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits="no-rounding",
+                ),
+            ),
+            NoaaDataVar(
+                name="categorical_ice_pellets_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="cicep",
+                    long_name="Categorical ice pellets",
+                    units="0=no; 1=yes",
+                    step_type="avg",
+                    comment="Presence/absence over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="CICEP",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=605,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits="no-rounding",
+                ),
+            ),
+            NoaaDataVar(
+                name="categorical_freezing_rain_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="cfrzr",
+                    long_name="Categorical freezing rain",
+                    units="0=no; 1=yes",
+                    step_type="avg",
+                    comment="Presence/absence over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="CFRZR",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=606,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits="no-rounding",
+                ),
+            ),
+            NoaaDataVar(
+                name="categorical_rain_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="crain",
+                    long_name="Categorical rain",
+                    units="0=no; 1=yes",
+                    step_type="avg",
+                    comment="Presence/absence over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="CRAIN",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=607,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits="no-rounding",
+                ),
+            ),
+            NoaaDataVar(
+                name="precipitable_water_atmosphere",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="pwat",
+                    long_name="Precipitable water",
+                    units="kg/(m^2)",
+                    step_type="instant",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="PWAT",
+                    grib_description='0[-] EATM="Entire atmosphere (considered as a single layer)"',
+                    grib_index_level="entire atmosphere (considered as a single layer)",
+                    index_position=625,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="total_cloud_cover_atmosphere",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="tcc",
+                    long_name="Total Cloud Cover",
+                    units="%",
+                    step_type="avg",
+                    comment="Average over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="TCDC",
+                    grib_description='0[-] EATM="Entire Atmosphere"',
+                    grib_index_level="entire atmosphere",
+                    index_position=635,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="geopotential_height_cloud_ceiling",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="gh",
+                    long_name="Geopotential height",
+                    units="gpm",
+                    step_type="instant",
+                    standard_name="geopotential_height",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="HGT",
+                    grib_description='0[-] CEIL="Cloud ceiling"',
+                    grib_index_level="cloud ceiling",
+                    index_position=637,
+                    keep_mantissa_bits=8,
+                ),
+            ),
+            NoaaDataVar(
+                name="downward_short_wave_radiation_flux_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="sdswrf",
+                    long_name="Surface downward short-wave radiation flux",
+                    units="W/(m^2)",
+                    step_type="avg",
+                    comment="Average over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="DSWRF",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=652,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="downward_long_wave_radiation_flux_surface",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="sdlwrf",
+                    long_name="Surface downward long-wave radiation flux",
+                    units="W/(m^2)",
+                    step_type="avg",
+                    comment="Average over the previous 1-6 hours, reset every 6-hour forecast step (00Z, 06Z, 12Z, 18Z).",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="DLWRF",
+                    grib_description='0[-] SFC="Ground or water surface"',
+                    grib_index_level="surface",
+                    index_position=653,
+                    window_reset_frequency=default_window_reset_frequency,
+                    keep_mantissa_bits=default_keep_mantissa_bits,
+                ),
+            ),
+            NoaaDataVar(
+                name="pressure_reduced_to_mean_sea_level",
+                encoding=encoding,
+                attrs=DataVarAttrs(
+                    short_name="prmsl",
+                    long_name="Pressure reduced to MSL",
+                    units="Pa",
+                    step_type="instant",
+                ),
+                internal_attrs=NoaaInternalAttrs(
+                    grib_element="PRMSL",
+                    grib_description='0[-] MSL="Mean sea level"',
+                    grib_index_level="mean sea level",
+                    index_position=0,
+                    keep_mantissa_bits=10,
+                ),
+            ),
+        ]
